Changes:  - fixed an object life-time issue that sometimes caused deadlocks
    Most of the time the tool is first to break connection, but if not there
    was the chance for this issue to occur. This time all locking was correct:
    locking order and relative positioning of lock statements. A disconnect
    operation could get blocked waiting for a write operation in progress. But
    the handler (to unblock the disconnect operation) for the write operation
    was never executed because the connection object was destroyed when the
    controller broke the connection.

 - fixed small process termination problems

    - sometimes processes where forcibly terminated more than once
    - sometimes a new process was forcibly terminated but the termination
      operation was meant to kill the process that went before, (the configuration process)

 - fixed small issue: the lower bound of the range of a progress bar was not
   being initialised (by default)



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@3665 21252fe5-60e8-0310-8ff0-c1ae1001594e

+++ tools/squadt/process.cpp
@@ -33,7 +33,7 @@
 #include <boost/format.hpp>
 #include <boost/thread/thread.hpp>
 
-#include "task_monitor.hpp"
+#include "task_monitor.tcc"
 #include "command.hpp"
 #include "mcrl2/setup.h"
 
@@ -139,7 +139,7 @@ namespace squadt {
     void process_impl::terminate() {
       /* Inform monitor */
       boost::shared_ptr < task_monitor > l = monitor.lock();
-      
+
       if (l.get() != 0) {
         l->disconnect(interface_pointer);
       }
@@ -160,7 +160,11 @@ namespace squadt {
         }
       }
 #else
-      kill(identifier, SIGKILL);
+      if (current_status == process::running) {
+        kill(identifier, SIGKILL);
+
+        current_status = process::aborted;
+      }
 #endif
     }
  
@@ -180,7 +184,7 @@ namespace squadt {
     
         boost::shared_ptr < process > alive(p.lock());
 
-        if (alive.get())  {
+        if (alive.get()) {
           boost::shared_ptr < process_impl >& impl(alive->impl);
 
           impl->current_status = (GetExitCodeProcess(ph, &exit_code) && exit_code == 0) ? process::completed : process::aborted;
@@ -203,7 +207,7 @@ namespace squadt {
 
       boost::shared_ptr < process > alive(p.lock());
 
-      if (alive.get())  {
+      if (alive.get()) {
         boost::shared_ptr < process_impl >& impl(alive->impl);
 
         impl->current_status = (WIFEXITED(exit_code)) ? process::completed : process::aborted;
@@ -220,7 +224,7 @@ namespace squadt {
       boost::shared_ptr < task_monitor > l = monitor.lock();
 
       if (l.get() != 0) {
-        l->signal_change(current_status);
+        l->signal_change(interface_pointer.lock(), current_status);
       }
     }
